backend/auth-service/requirements.txt
fastapi
uvicorn
email-validator
motor
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]

